Map eingebunden

+++ Regensbad/app/src/main/java/com/example/tom/regensbad/Domain/CivicPool.java
@@ -14,11 +14,11 @@ public class CivicPool {
     private double longi;
     private String phoneNumber;
     private String website;
-    private double openTime;
-    private double closeTime;
+    private int openTime;
+    private int closeTime;
     private String picPath;
 
-    public CivicPool (String name ,String type, double lati, double longi, String phoneNumber, String website, double openTime, double closeTime, String picPath){
+    public CivicPool (String name ,String type, double lati, double longi, String phoneNumber, String website, int openTime, int closeTime, String picPath){
         this.name = name;
         this.type = type;
         this.lati = lati;
@@ -64,11 +64,11 @@ public class CivicPool {
         return webstieString;
     }
 
-    public double getOpenTime(){
+    public int getOpenTime(){
         return openTime;
     }
 
-    public double getCloseTime(){
+    public int getCloseTime(){
         return closeTime;
     }
